💡 chore : black을 이용한 따옴표 설정 통일

+++ backend/app/wishlist/tests.py
@@ -42,31 +42,23 @@ class WishlistTestCase(APITestCase):
         # 위시리스트에 처음 넣는 경우 테스트
         url = reverse("wishlist")
         data = {"product_id": self.product1.id}
-        response = self.client.post(
-            url, data, headers={"Authorization": f"Bearer {self.token}"}
-        )
+        response = self.client.post(url, data, headers={"Authorization": f"Bearer {self.token}"})
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
         self.assertTrue(response.data["status"])
         self.assertEqual(response.data["product"], self.product1.id)
         self.assertEqual(response.data["user"], self.user.id)
 
         # 이미 위시리스트에 있는 경우 테스트
-        response = self.client.post(
-            url, data, headers={"Authorization": f"Bearer {self.token}"}
-        )
+        response = self.client.post(url, data, headers={"Authorization": f"Bearer {self.token}"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual("already added", response.data["msg"])
 
         # 이미 위시리스트에 넣었다가 취소하고 다시 위시리스트에 추가하는 경우 테스트
-        wishlist = Wishlist.objects.get(
-            user_id=self.user.id, product_id=self.product1.id
-        )
+        wishlist = Wishlist.objects.get(user_id=self.user.id, product_id=self.product1.id)
         wishlist.status = False
         wishlist.save()
-        response = self.client.post(
-            url, data, headers={"Authorization": f"Bearer {self.token}"}
-        )
+        response = self.client.post(url, data, headers={"Authorization": f"Bearer {self.token}"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertTrue(response.data["status"])
@@ -80,9 +72,7 @@ class WishlistTestCase(APITestCase):
 
         url = reverse("wishlist")
 
-        response = self.client.get(
-            url, headers={"Authorization": f"Bearer {self.token}"}
-        )
+        response = self.client.get(url, headers={"Authorization": f"Bearer {self.token}"})
         pdb.set_trace()
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(len(response.data), 3)
@@ -119,9 +109,7 @@ class WishlistDetailTestCase(APITestCase):
 
     def test_get_wishlist_detail(self):
         url = reverse("wishlist_detail", kwargs={"product_id": self.product.id})
-        response = self.client.get(
-            url, headers={"Authorization": f"Bearer {self.token}"}
-        )
+        response = self.client.get(url, headers={"Authorization": f"Bearer {self.token}"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.data["product"]["id"], self.product.id)
@@ -131,34 +119,24 @@ class WishlistDetailTestCase(APITestCase):
         self.assertEqual(response.data["user"], self.user.id)
 
         url = reverse("wishlist_detail", kwargs={"product_id": 919919191911})
-        response = self.client.get(
-            url, headers={"Authorization": f"Bearer {self.token}"}
-        )
+        response = self.client.get(url, headers={"Authorization": f"Bearer {self.token}"})
 
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
     def test_delete_product_detail(self):
         url = reverse("wishlist_detail", kwargs={"product_id": self.product.id})
-        response = self.client.delete(
-            url, headers={"Authorization": f"Bearer {self.token}"}
-        )
-        deleted_wishlist = Wishlist.objects.get(
-            user_id=self.user.id, product_id=self.product.id
-        )
+        response = self.client.delete(url, headers={"Authorization": f"Bearer {self.token}"})
+        deleted_wishlist = Wishlist.objects.get(user_id=self.user.id, product_id=self.product.id)
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertFalse(deleted_wishlist.status)
 
-        response = self.client.delete(
-            url, headers={"Authorization": f"Bearer {self.token}"}
-        )
+        response = self.client.delete(url, headers={"Authorization": f"Bearer {self.token}"})
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.data["msg"], "already deleted")
 
         url = reverse("wishlist_detail", kwargs={"product_id": 919919191911})
-        response = self.client.get(
-            url, headers={"Authorization": f"Bearer {self.token}"}
-        )
+        response = self.client.get(url, headers={"Authorization": f"Bearer {self.token}"})
 
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
